Extract demo logic into public example modules
Move DemoApp and camera_sim logic from bin/ into
src/examples/ so downstream consumers can reuse
them without duplicating code.
