+++ crates/sophus/src/examples.rs
@@ -1,3 +1,8 @@
+/// Offscreen camera simulator example.
+#[cfg(feature = "std")]
+pub mod camera_sim;
+/// Demo application with optics simulation and 3D viewer.
+pub mod demo_app;
 /// Optics simulation components.
 pub mod optics_sim;
 /// Viewer example utilities.
